fix: Cannot build for aws lambda
fix: https://github.com/appleboy/gorush/issues/432
